ab_executor_slots

Struct Slots

pub struct Slots(/* private fields */);

Collection of slots, primarily for the execution environment

Implementations

impl Slots

pub fn new<I>(slots: I) -> Self
where I: IntoIterator<Item = Slot>,

Create a new instance from a hashmap containing existing slots.

Only slots that are present in the input can be modified. The only exception is slots for owners created during runtime and initialized with Self::add_new_contract().

โEmptyโ slots must still have a value in the form of an empty [SharedAlignedBuffer].

pub fn new_nested_rw(&mut self) -> NestedSlots<'_>

Create a new read-write NestedSlots instance.

Nested instance will integrate its changes into the parent slot when dropped (or changes can be reset with NestedSlots::reset()).

pub fn new_nested_ro(&self) -> NestedSlots<'_>

Create a new read-only NestedSlots instance

pub fn add_new_contract(&mut self, owner: Address) -> bool

Add a new contract that didnโt exist before.

In contrast to contracts in Self::new(), this contract will be allowed to have any slots related to it being modified.

Returns false if a contract already exits in a map, which is also considered as an access violation.

pub fn iter( &self, ) -> impl ExactSizeIterator<Item = (&SlotKey, &SharedAlignedBuffer)> + '_

Iterate over all slots in the collection

pub fn iter_modified( &self, ) -> impl Iterator<Item = (&SlotKey, &SharedAlignedBuffer)> + '_

Iterate over modified slots in the collection

pub fn into_slots( self, ) -> impl ExactSizeIterator<Item = (SlotKey, SharedAlignedBuffer)>

Extract all slots in the collection